% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aatoss.R
\name{getResultsFromServer}
\alias{getResultsFromServer}
\title{Get results from server MUFOLD server}
\usage{
getResultsFromServer(send_object)
}
\arguments{
\item{send_object}{object which was obtained by function 'sendToServer'}
}
\value{
data.frame with prediction of secondary structure
}
\description{
Get results from server MUFOLD server
}
\examples{
email = "test11@test.com"
testseq <- data.frame(target = c("fasta_seq1", "fasta_seq2"),
                      fasta  = c("RDPHPAPPRTSQEPHQNPHGVIPSESKPFVASKPKPHT",
                                 "PSLPRPPGCPAQQGGSAPIDPPPVHESPHPPLPATEPASRLSSE"),
                      stringsAsFactors=FALSE)

# send sequences to server

sequences <- sendToServer(testseq, email)

# wait a couple of minutes, 1 sequence ~ 1-2 minutes of calculation

# retrive results from server ( only finished sequences )

structures <- getResultsFromServer(sequences)

# or you can obtain all finished sequences which associated with this email

all.structures <- getAllResultsFromServer(email)
}
