% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aatoss.R
\name{sendToServer}
\alias{sendToServer}
\title{Title Send sequences to server}
\usage{
sendToServer(df_sequences, email)
}
\arguments{
\item{df_sequences}{data.frame with sequences (columns target and fasta required)}

\item{email}{at the moment any email-like identifier, it is not checked}
}
\value{
object which you can use as input to retrieve results from server
}
\description{
Title Send sequences to server
}
\examples{
email = "test11@test.com"
testseq <- data.frame(target = c("fasta_seq1", "fasta_seq2"),
                      fasta  = c("RDPHPAPPRTSQEPHQNPHGVIPSESKPFVASKPKPHT",
                                 "PSLPRPPGCPAQQGGSAPIDPPPVHESPHPPLPATEPASRLSSE"),
                      stringsAsFactors=FALSE)

# send sequences to server

sequences <- sendToServer(testseq, email)

# wait a couple of minutes, 1 sequence ~ 1-2 minutes of calculation

# retrive results from server ( only finished sequences )

structures <- getResultsFromServer(sequences)

# or you can obtain all finished sequences which associated with this email

all.structures <- getAllResultsFromServer(email)
}
